model/requirements.txt
numpy
pandas
scikit-learn
scipy
flask
gunicorn
torch
torchvision
transformers
faiss-cpu
kafka-python
joblib
requests
tqdm
scikit-surprise